requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
flask
river
polars
gdown
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
-e .